pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "crminaec.__about__.__version__"}

[project]
name = "crminaec"  # Updated to match the package name
dynamic = ["version"]
description = "Interoperable CRM, AEC & Education Management Ecosystem"
readme = "README.md"
authors = [
    {name = "Dr. Kerem ERCOŞKUN", email = "ercoskunkerem@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"

dependencies = [
    "Flask>=3.0.0",
    "Flask-Login>=0.6.3",     # 🚨 CRITICAL: Required for the Traffic Cop
    "Flask-SQLAlchemy>=3.1.0",
    "SQLAlchemy>=2.0.0",
    "Flask-WTF>=1.1.0",
    "python-dotenv>=1.0.0",
    "waitress>=3.0.0",        # Required for IIS deployment
    "markdown2>=2.4.0",
    "jsTree>=0.0.1",          # If you use the python wrapper, otherwise handled in JS
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "cryptography>=40.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "coverage>=7.0.0",
    "mypy>=1.5.0",
    "pylint>=2.17.0",
]
api = [
    "Flask-RESTful>=0.3.10",
    "Flask-CORS>=4.0.0",
]
reports = [
    "weasyprint>=58.0",
    "pdfkit>=1.0.0",
    "markdown2>=2.4.0",
]
interop = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/keremer/pearson"
Documentation = "https://github.com/keremer/pearson/wiki"

[project.scripts]
# Updated to point to the main module within the pearson package
crminaec = "crminaec.main:cli"

[tool.setuptools.packages.find]
where = ["."]
# Focus on the 'crminaec' root package; setuptools will find the 
# sub-packages (web, api, cli, interop) automatically if they contain __init__.py
include = ["crminaec*"]
exclude = ["tests*", "*.tests", "*.tests.*", "tests.*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true